requirements.txt
Django==1.8.2
django-cors-headers==1.1.0
djangorestframework==3.1.2
djangorestframework-camel-case==0.2.0
wheel==0.24.0
